// packages/emails/email-types.ts
export enum EmailType {
  CONFIRMATION = "confirmation",
  CANCELLATION = "cancellation",
  RESCHEDULED = "rescheduled",
  REQUEST = "request",
  REASSIGNED = "reassigned",
  AWAITING_PAYMENT = "awaiting_payment",
  RESCHEDULE_REQUEST = "reschedule_request",
  LOCATION_CHANGE = "location_change",
  NEW_EVENT = "new_event",
}